moved Error and support into error.rs
