+++ src/lib.rs
@@ -1,6 +1,8 @@
 extern crate flate2;
 extern crate byteorder;
 
+pub use error::Error;
 pub use nbt::Tag;
 
+pub mod error;
 pub mod nbt;
